+++ src/qlext/QLBridging.h
@@ -1,0 +1,2 @@
+#include "cadmesh.h"
+#include "archivelist.h"